LPS-151028 Merge BaseConnectionPoolMetrics into HikariConnectionPoolMetrics
