+++ portal-impl/src/com/liferay/portal/dao/jdbc/pool/metrics/HikariConnectionPoolMetrics.java
@@ -14,14 +14,24 @@
 
 package com.liferay.portal.dao.jdbc.pool.metrics;
 
+import com.liferay.portal.dao.jdbc.aop.DefaultDynamicDataSourceTargetSource;
+import com.liferay.portal.kernel.bean.PortalBeanLocatorUtil;
+import com.liferay.portal.kernel.dao.jdbc.pool.metrics.ConnectionPoolMetrics;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+
 import com.zaxxer.hikari.HikariPoolMXBean;
 
 import java.lang.reflect.Method;
 
+import org.springframework.aop.framework.Advised;
+import org.springframework.aop.support.AopUtils;
+import org.springframework.jdbc.datasource.LazyConnectionDataSourceProxy;
+
 /**
  * @author Mladen Cikara
  */
-public class HikariConnectionPoolMetrics extends BaseConnectionPoolMetrics {
+public class HikariConnectionPoolMetrics implements ConnectionPoolMetrics {
 
 	public HikariConnectionPoolMetrics(Object dataSource)
 		throws ReflectiveOperationException {
@@ -33,6 +43,15 @@ public class HikariConnectionPoolMetrics extends BaseConnectionPoolMetrics {
 		_getPoolNameMethod = clazz.getMethod("getPoolName");
 
 		_getHikariPoolMXBeanMethod = clazz.getMethod("getHikariPoolMXBean");
+	}
+
+	@Override
+	public String getConnectionPoolName() {
+		if (_name == null) {
+			_name = _getConnectionPoolName();
+		}
+
+		return _name;
 	}
 
 	@Override
@@ -63,13 +82,65 @@ public class HikariConnectionPoolMetrics extends BaseConnectionPoolMetrics {
 		}
 	}
 
-	@Override
-	protected Object getDataSource() {
-		return _dataSource;
-	}
+	private String _getConnectionPoolName() {
+		LazyConnectionDataSourceProxy lazyConnectionDataSourceProxy =
+			(LazyConnectionDataSourceProxy)PortalBeanLocatorUtil.locate(
+				"counterDataSource");
 
-	@Override
-	protected String getPoolName() {
+		if (_dataSource.equals(
+				lazyConnectionDataSourceProxy.getTargetDataSource())) {
+
+			return "counterDataSource";
+		}
+
+		lazyConnectionDataSourceProxy =
+			(LazyConnectionDataSourceProxy)PortalBeanLocatorUtil.locate(
+				"liferayDataSource");
+
+		Object targetDataSource =
+			lazyConnectionDataSourceProxy.getTargetDataSource();
+
+		if (_dataSource.equals(targetDataSource)) {
+			return "liferayDataSource";
+		}
+		else if (AopUtils.isAopProxy(targetDataSource) &&
+				 (targetDataSource instanceof Advised)) {
+
+			Advised advised = (Advised)targetDataSource;
+
+			targetDataSource = advised.getTargetSource();
+
+			if (targetDataSource instanceof
+					DefaultDynamicDataSourceTargetSource) {
+
+				try {
+					DefaultDynamicDataSourceTargetSource
+						defaultDynamicDataSourceTargetSource =
+							(DefaultDynamicDataSourceTargetSource)
+								targetDataSource;
+
+					if (_dataSource.equals(
+							defaultDynamicDataSourceTargetSource.
+								getReadDataSource())) {
+
+						return "readDataSource";
+					}
+
+					if (_dataSource.equals(
+							defaultDynamicDataSourceTargetSource.
+								getWriteDataSource())) {
+
+						return "writeDataSource";
+					}
+				}
+				catch (Exception exception) {
+					if (_log.isDebugEnabled()) {
+						_log.debug(exception);
+					}
+				}
+			}
+		}
+
 		try {
 			return (String)_getPoolNameMethod.invoke(_dataSource);
 		}
@@ -78,8 +149,12 @@ public class HikariConnectionPoolMetrics extends BaseConnectionPoolMetrics {
 		}
 	}
 
+	private static final Log _log = LogFactoryUtil.getLog(
+		HikariConnectionPoolMetrics.class);
+
 	private final Object _dataSource;
 	private final Method _getHikariPoolMXBeanMethod;
 	private final Method _getPoolNameMethod;
+	private String _name;
 
 }
